requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.13.3
gunicorn==21.2.0
